chore(release): Disable NPM release temporarily
This commit disables the release of a NPM package for the 2.0.0 release. 
This is because a past failure from semantic-release has caused the 
package to be published on NPM but not on Github. As we have fixed the 
semantic-release/github issue, we also unpublished the NPM package. We 
now publish a 2.0.0 version via semantic-release for the sake of it, and 
will immediately release a minor update with a new feature to provide a 
valid NPM package with a version above 2.0.0.